refactoring
better way to read in the information from the configuration file
